Single Case data processing
I need to rewrite the process so it handles for the ngram case. instead
of the single gram case

Shit I need to go back and generalise a few things

Make it so the cleaning works for ngrams. Single and multiple cases.
The text needs to be cleaned first before creating the ngrams
Once the ngrams exist then have a standard function t perform the
analysis on each of the datasets

I need to take a step back and make a concept of how the data flows. And
then implement each step as an independant block

Read the data
Clean the data
Process data into ngrams
Analyse data